+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # master
 - Verify Password command
 - Email Auth command
+- Domain Owner command
 - Better tests
 
 # 0.1.0 (2017-04-15)